requirements.txt
click~=8.1
PyYAML==6.0
